requirements.txt
fastapi
uvicorn
pydantic
passlib[bcrypt]
python-jose
mysql-connector-python
python-multipart
numpy
Pillow
tensorflow
pyjwt